requirements.txt
tk
Pillow==8.2.0
numpy==1.19.5
opencv-python
